Stylized and partially responsive web dashboard
This sample web dashboard was built as an exercise using
CSS Grid to style and organize a web layout. The layout contains a
navigation sidebar, header with search bar and other currently
non-functional buttons, and a main content section with two parts.
The first part of the main content section is a grid of two columns,
containing content items consisting of a header, text, and 3 svg-styled
<a> elements. The right side of the main content section is meant to model
a sidebar for announcements as well as other suggested users.

The layout is not currently designed with mobile in mind, but all the
sections of the layout have built in responsiveness to allow for some
various screen sizes. All buttons and <a> elements are currently
non-functional, but should be ready to be connected to their appropriate
back-end links.

This project was built almost entirely using CSS Grid, although Flex was
used in two places for ease of styling.